fx 6,9

+++ 0x09-csharp-linear_algebra/6-vector_addition/6-vector_addition.cs
@@ -1,6 +1,9 @@
 ﻿using System;
+
+/// <summary>This is the class object.</summary>
 class VectorMath
 {
+    /// <summary>This is the function</summary>
     public static double[] Add(double[] vector1, double[] vector2){
         double[] add = new double[2];
         
@@ -11,7 +14,7 @@ class VectorMath
             return add;
         }
         
-        double[] result = new double[(int) vector1.Length];
+        double[] result = new double[vector1.Length];
         
         for (int i = 0; i < vector1.Length; i++){
             result[i] = vector1[i] + vector2[i];
